+++ mordant/src/commonMain/kotlin/com/github/ajalt/mordant/animation/Animation.kt
@@ -7,22 +7,25 @@ import com.github.ajalt.mordant.widgets.RawWidget
 import com.github.ajalt.mordant.widgets.Text
 
 /**
- * An Animation renders a widget to the screen each time [update] is called, clearing the render from the previous call.
+ * An Animation renders a widget to the screen each time [update] is called, clearing the render
+ * from the previous call.
  *
- * If you [print][Terminal.print] anything in between calls to [update], the rendered widget will be cleared, before the
- * print, then rendered again below it.
+ * If you [print][Terminal.print] anything in between calls to [update], the rendered widget will be
+ * cleared, before the print, then rendered again below it.
  *
- * Note that [Animation] doesn't render the widget automatically: you need to call [update] yourself when your data
- * changes. If your terminal is not [interactive][TerminalInfo.interactive], the animation will not render anything.
+ * Note that [Animation] doesn't render the widget automatically: you need to call [update] yourself
+ * when your data changes. If your terminal is not [interactive][TerminalInfo.interactive], the
+ * animation will not render anything.
  *
- * You can create instances of Animations with [animation], [textAnimation], and `progressAnimation` (on JVM), or by creating a
- * subclass.
+ * You can create instances of Animations with [animation], [textAnimation], and `progressAnimation`
+ * (on JVM), or by creating a subclass.
  */
 @OptIn(ExperimentalTerminalApi::class)
 abstract class Animation<T>(private val terminal: Terminal) {
     private var size: Pair<Int, Int>? = null
     private var text: String? = null
     private var needsClear = false
+    private var interceptorInstalled = false
 
     // Don't move the cursor the first time the animation is drawn
     private var firstDraw = true
@@ -30,12 +33,8 @@ abstract class Animation<T>(private val terminal: Terminal) {
     private val interceptor: TerminalInterceptor = TerminalInterceptor { req ->
         text?.let { t ->
             PrintRequest(text = buildString {
-                if (req.text.isNotEmpty()) {
-                    needsClear = true
-                }
-
                 if (!firstDraw) {
-                    getClear()?.let { append(it) }
+                    getCursorMoves(needsClear || req.text.isNotEmpty())?.let { append(it) }
                 }
                 firstDraw = false
                 if (req.text.isNotEmpty()) {
@@ -46,33 +45,45 @@ abstract class Animation<T>(private val terminal: Terminal) {
         } ?: req
     }
 
-    init {
-        if (terminal.info.outputInteractive) {
-            terminal.addInterceptor(interceptor)
-        }
-    }
-
     protected abstract fun renderData(data: T): Widget
 
+    /**
+     * Stop this animation and remove it from the screen.
+     *
+     * Future calls to [update] will cause the animation to resume.
+     */
     fun clear() {
-        needsClear = true
-        getClear()?.let {
-            text = null
-            terminal.removeInterceptor(interceptor)
-            terminal.print(RawWidget(it))
-        }
+        stop()
+        getCursorMoves(clearScreen = true)?.let { terminal.print(RawWidget(it)) }
+        size = null
     }
 
-    private fun getClear(): String? {
-        val (height, _) = size ?: return null
-        return terminal.cursor.getMoves {
-            startOfLine()
-            up(height)
-            if (needsClear) clearScreenAfterCursor()
-        }
+    /**
+     * Stop this animation without removing it from the screen.
+     *
+     * Anything printed to the terminal after this call will be printed below this last frame of
+     * this animation.
+     *
+     * Future calls to [update] will cause the animation to start again.
+     */
+    fun stop() {
+        if (interceptorInstalled) terminal.removeInterceptor(interceptor)
+        interceptorInstalled = false
+        firstDraw = true
+        text = null
     }
 
+    /**
+     * Update the data for this animation.
+     *
+     * This will cause the existing animation frame to be cleared and the new frame rendered in its
+     * place.
+     */
     fun update(data: T) {
+        if (!interceptorInstalled && terminal.info.outputInteractive) {
+            terminal.addInterceptor(interceptor)
+        }
+        interceptorInstalled = true
         val rendered = renderData(data).render(terminal)
         val height = rendered.height
         val width = rendered.width
@@ -82,6 +93,15 @@ abstract class Animation<T>(private val terminal: Terminal) {
         text = terminal.render(rendered)
         // Print an empty renderable to trigger our interceptor, which will add the rendered text
         terminal.print(EmptyWidget)
+    }
+
+    private fun getCursorMoves(clearScreen: Boolean): String? {
+        val (height, _) = size ?: return null
+        return terminal.cursor.getMoves {
+            startOfLine()
+            up(height)
+            if (clearScreen) clearScreenAfterCursor()
+        }
     }
 }